refactor: extract TM analysis models
